Separated classes in two files and created sample script
